refactor: reorganiza a estrutura do projeto criando pasta de loaders
