+++ src/main.ts
@@ -1,21 +1,22 @@
 import "reflect-metadata";
-import dotenv from "dotenv";
-import { app } from "./app";
+import * as dotenv from "dotenv";
 
 import { StringColors, StringUtils } from "@utils/string";
-import AppDataSource from "@services/database";
+
+import App from "./loaders/app";
+import AppDataSource from "./loaders/database";
+import routes from "./routes";
 
 dotenv.config();
 
-const start = async () => {
+export const start = async () => {
   try {
     await AppDataSource.initialize();
-
-    console.log(`${StringUtils.app} | Banco de dados conectado!`);
-
-    await AppDataSource.synchronize();
-
     console.log(`${StringUtils.app} | Banco de dados sincronizado!`);
+
+    const app = App().app;
+
+    app.use("/v1", routes);
 
     app.listen(process.env.PORT || 3000, () => {
       console.log(`${StringUtils.app} | O Servidor foi iniciado: ${StringColors.yellow}http://localhost:${process.env.PORT || 3000}${StringColors.null}`);
