+++ src/repositores/index.ts
@@ -1,6 +1,6 @@
 import Record from "@modules/Record/record.model";
 import User from "@modules/User/user.model";
-import AppDataSource from "@services/database";
+import AppDataSource from "src/loaders/database";
 
 export const UserRepository = () => AppDataSource.getRepository(User);
